Try content width column as new layout type
