+++ packages/block-editor/src/layouts/index.js
@@ -3,8 +3,9 @@
  */
 import flex from './flex';
 import flow from './flow';
+import column from './column';
 
-const layoutTypes = [ flow, flex ];
+const layoutTypes = [ flow, flex, column ];
 
 /**
  * Retrieves a layout type by name.
