+++ packages/block-library/src/group/block.json
@@ -66,7 +66,11 @@
 				"fontSize": true
 			}
 		},
-		"__experimentalLayout": true
+		"__experimentalLayout": {
+			"default": {
+				"type": "column"
+			}
+		}
 	},
 	"editorStyle": "wp-block-group-editor",
 	"style": "wp-block-group"
